refactor(oneseo): EvaluationStatus enum class 위치를 application.type 으로 변경
