+++ src/main/java/team/themoment/hellogsmv3/domain/application/dto/request/ApplicationStatusReqDto.java
@@ -2,7 +2,7 @@ package team.themoment.hellogsmv3.domain.application.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import team.themoment.hellogsmv3.domain.oneseo.entity.type.EvaluationStatus;
+import team.themoment.hellogsmv3.domain.application.type.EvaluationStatus;
 import team.themoment.hellogsmv3.domain.oneseo.entity.type.Major;
 import team.themoment.hellogsmv3.domain.oneseo.entity.type.Screening;
