+++ src/main/java/team/themoment/hellogsmv3/domain/application/type/EvaluationStatus.java
@@ -1,4 +1,4 @@
-package team.themoment.hellogsmv3.domain.oneseo.entity.type;
+package team.themoment.hellogsmv3.domain.application.type;
 
 public enum EvaluationStatus {
     NOT_YET,
